Only lock the db during indexing.
Previously we used to keep a lock on the leveldb throughout the full
lifetime of the server. This was unneccessary, since we only use the
lock for a brief period during indexing. This change should minimize
locking issues, but it does not solve them completely. To solve this
permanently, we should add a retry mechanism.