Adjust position of indexed definitions.
Previously, we only used TokenEditDistanceTest to find the correct
symbol but we did not adjust the positions returned from the server.
This resulted in bogus results in go to definition when the definition
appeared in the same document, for example a local symbol.

Mapping the result position turned out to be more difficult than I
expected because of how SymbolIndex is organized around
references/definitions and the fact that I don't want to tax the
common case when original == revised.

This commit refactors the signatures of TokenEditDistanceTest to
return a ternary option: `Either[NoMatch | Unchanged, Position]`.
The old `Option[Position]` type was not able to distinguish between
Unchanged and NoMatch.