Implement codeAction to remove unused imports with scalafix
Users can put the cursor or hover over an unused import to see a
lightbulb appear. If the user clicks on the lightbulb, a dialogue
appears with options to run commands like "Remove unused import".

This commit turned out to be quite tricky since I wasn't sure how to
submit a request from the server to the client for workspace/applyEdit.
It turned out to be quite dead simple, just send a JsonRpcRequestMessage
and discard the response.