Add benchmarking infrastructure
We care equally about memory useage and CPU time to index sources.
Rough summary of the results:

- indexing 7 million loc Java and 1.5 million loc Scala takes ~15M ram
- Scala topelevel parser handles ~800k loc/s

The numbers are so good that I'm not too stressed about
micro-optimization as the moment. We run the full indexer for a
medium-sized classpath on every unit test without annoying me,
total runtime of `unit/test` is ~4 seconds.