Fix deadlock in ConcurrentHashMap
Interestingly, `ConcurrentHashMap` freezes in one of the tests while
`Collections.synchronizedMap` works fine. My guess at why this happens
is that `ConcurrentHashMap` should only be used when the "if absent" function
is super fast and simple, whereas in InteractiveSemanticdbs the method
takes at least 100-200ms to construct a compiler instance.