Make Java version configurable in VS Code
Previously, the VS Code plugin ignored the `JAVA_HOME` environment
variable and bindly used Java 11 if it was installed on the machine
resulting in cryptic errors. Now VS Code respects `JAVA_HOME` and users
can override the location through settings.

If the server is started in an unsupported Java, then Metals now reports
a helpful error message instead of crashing with a cryptic stack trace.