Improve compilation tracking in the status bar.
Previously, we showed a single "Compiling $project" message when
compiling a project regardless if that project triggered compilation
in other projects in the workspace. With this commit, we now add
a "Compiling $project" status bar message for each unique project
that's compiled.

For example, if you open `core-test` with one transitive dependency you
get two status bar updates.

- "Compiling core": the main sources.
- "Compiling core-test": the test sources.


Former-commit-id: 3a8346df0127d50e8e9c6addc69333813fa69219