Undo IntelliJ Java rename.
I had renamed a Java interface method name from `name` to `displayName`
and it seems the default behavior is to update all occurrences of "name"
anywhere in the codebase (comments, string literals) to "displayName".
I had to rush out of the door so I committed everything in a rush
and didn't notice until the tests started failing (thankfully!).