Handle generic unapply signatures in signature help.
Previously, we showed `unapply[A, B](x$1: Left[A, B]): Option[A]`
and now we show `unapply(value: A): Some[A]` where `value: A` is the
field on the `Left` case class.

We can still do much better for custom extractors but that can be left
for a separate PR. This is a bugfix.