Remove signature help fallback in hover.
Previously, hover used signature help as a fallback when the cursor was
over a named argument. After this commit, we show the method signature
instead. It would be nice to show only the parameter signature in the
future but this is commit is just to fix a bug that resulted in wrong
results.