Eagerly load presentation compiler for open buffers.
Previously, we waited until the first completion/hover/signatureHelp
request to load the presentation compiler. After this commit, we load
the compiler eagerly and display a "slow task" status with timer so that
it's more clear that we're downloading artifacts from the internet.
Scalafmt also uses slow tasks now, for consistency.