Don't eagerly load presentation compiler during tests.
Previously, we always eagerly loaded the presentation compiler to
improve the out-of-the-box experience for new Metals users. However,
this behavior is undesirable for tests because it triggers a lot
of unnecessary computation on every test.

It's generally undesirable to customize the runtime for running tests
but in this particular case I feel it's justified.