Make complications work on overr<COMPLETE> and def<COMPLETE>.
Fixes #631. Previously, we needed to write def <COMPLETE> to autocomplete an override, now we can write a start of the 'override' word or a combination of letters in the def/val name and it will trigger the proper completion. When starting with override word it will add it to method even if overriding abstract methods.

Also fixes a bug with not autocompleting when overriding vals form traits.