Clear diagnostics on build import, fixes #644.
Previously, we could occasionally get in a state where Metals published
stale diagnostics from an old build server.  Now, when we disconnect
from a build server we unpublish the diagnostics from that server. If we
connect to a new server, the errors will get re-published if they are
still relevant.