Setup git hooks for scalafmt and scalafix.
Previously, contributors needed to manually run scalafmt and scalafix
when opening pull requests in order to make the CI green. Now, we
automatically install git hooks so that scalafmt and scalafix are always
run before pushing a commit.

To speed up the `sbt scalafixAll` command, we use the experimental
`-client` flag which delegates the command to a running sbt server
instance if one exists. In my local experiments, this makes the pre-push
hooks run in ~5 seconds if the project is already compiled.