Fix CompletionKeywordSuite remaining tests
In some instances, keywords completions fallback to parsing the
diagnostics reported by the parser to infer the cursor position.

In 2.13 there have been some changes to the concrete implementations of
the Reporter class. Namely, the concrete class in 2.13 is
Reporter.AdaptedReporter, which delegates to a StoreReporter (whereas
in 2.12 we get a StoreReporter directly)

The storeReporter method provides a cross-compatible API for accessing
the StoreReporter instance