Only add Scala version prefix to tests in CI.
Previously, we always added the Scala version to test names when the
Scala version was not 2.12 in order to make it easier to track test
failures in CI logs for non-2.12 version. Now, we only do this test name
change in CI so that the tests have the same name when testing locally.