Fix off-by-one for parameter hints.
Previously, running parameter hints after the closing parenthesis in
`List(1)@@` or between curried arguments `Option(1).fold(1)@@(_ + 1)`
would trigger a false-positive signature help (resulting in noisy
popups). Now, we exclude the last characters such as closing `)` and
`]` to avoid returning parameter hints when the cursor is not inside
the argument position.

Thanks to Sébastien Doeraene for catching this bug!