Make code navigation work for any readonly/ file.
Previously, we relied on carefully recording what build target produced
what files in the `.metals/readonly/` directory. This approach had the
problem that navigation failed to work in `readonly/` sources if

- a new metals feature forgot to record the build target
- a user removes `.metals/metals.h2.db`

Now, when encountering an unknown `readonly/` file we do the following
steps to infer what build target it belongs to:

- extract toplevel symbol definitions from the source code.
- find a jar file from any classfile that defines one of the toplevel
  symbols.
- find the build target which has that jar file in it's classpath.

This approach is not glamorous but it seems to work reasonably well.