Add support for the Pants build tool.
Previously, it was not possible to use Metals with the Pants build tool.
This commit implements all the functionality to use Metals with Pants,
along with a lot smaller fixes that are necessary to work with Metals in
a larger workspace (encountered while developing the Pants integration).

This is commit is large because it was difficult to rebase the
individual commits on top of the Metals master.