Give Bloop more visibility on the website.
Previously, the website made it appear as if Bloop was an implementation
detail of Metals. Now, the website tries to give Bloop more visibility
and advertise it as an essential part of the Metals user experience (low
latency diagnostics, testing, running, soon debugging).

This change doesn't mean that Metals doesn't work with other BSP
servers. Given that >99% of Metals users use Bloop as the build server,
the new documentation only tries to make it clearer to users how
Metals work and how they can make the most of it.