Harden worksheet evaluation in the face of `Thread.stop()`.
Previously, when cancelling a worksheet we could kill the evaluation
thread with  `Thread.stop()` and prevent subsequent worksheets from
evaluating. I haven't been able to reproduce this situation with a test
case locally, but it looks the CI hit on this case.

Now, the `CompilerJobQueue` automatically restarts the single thread
executor if it's no longer running. This should fix the situation where
`Thread.stop()` prevents subsequent requests from evaluating.