Avoid race conditions when asking the user to import a build.
Previously, a race condiition could cause Metals to prompt the user
twice to import build changes. The duplicate prompt could cause some
test suites to become flaky. Now, we synchronize on the method that
prompts the user to ensure that we finish writing the digest
notification status in the H2 database before sending the next prompt.