Use Pants source directory instead of source file when possible.
Previously, when exporting Pants targets to Bloop, we always listed out
individual source files for each target. This approach had the problem
that we needed to update the Bloop JSON files when a new file got
created. Now, for Pants targets that use the recursive `**/*.scala`
file glob, we use source directories instead of source files making it
possible for bloop to catch new files as they are created.