Move resources to classpath.
Previously, we incorrectly used the "resources" field in the Bloop JSON config.
The "resources" field is only supposed to contain resources that are
defined in that project. We included all transitive dependencies
resources at every Bloop project, resulting in duplicate resource
entries.

Now, we add the resources to the classpath entry instead, which is
interpreted as "transitive resources".