Give constructors to InitializationOptions and ClientExperimentalCapabilities.
By giving primary constructors to these we are getting around the odd
behavior that gson has of setting missing keys to null or if an empty
`{}` is sent in from the client, we get a instance of the class with all
nulls. This protects against that and provides the actual defaults that
we want.