Add in check for JsonNull.
At the surface this commit makes no sense, but at times when editors are
leaving out the `InitializationOptions` key, they are somehow being
treated as a different type of `null`. One `null` is the `null` you'd
expect and is not matching as a `JsonElement` and then getting the
Default while the other _is_ matching `JsonElement`, so we add in a case
for `JsonNull`.