teach metals to accept configuration of range formatting
This PR provides an escape hatch to Emacs users who find the beahvior documented in
formatting can opt-in to this behavior.

My main concern is that this change appears not backwards compatible. I am not familiar
with Metals JSON deserialization system. Ideally the `InitializationOptions` would
default to `true` for `provideRangeFormater` if it is not present in the JSON payload
from the `initialize` request.

I will make a corresponding change to https://github.com/emacs-lsp/lsp-mode to support
this new configuration option.

Thank you for your time.