Add in a preparePr alias.
The idea behind this is to simplify the process for a contributor
to prepare their code for a pull-request. Currently a person has to
both execute the bin/scalafmt and also run scalafix via sbt. This
simplifies this to one step so a contributor can just run
`sbt preparePr` to run scalfix and then scalafmt.