Ensure the edit positions are correct for autoImports with Ammonite.
Previously the edit for autoImports when triggered from a code action would contain
reference to the position the presentation compiler had for the generated Scala
file that the Ammonite script created. Like other functionality that we have that
interacts with the Presentation compiler while dealing with Ammonite we ensure that
the edits have positions that are adjusted to ensure they land in the correct place
in the script, not the generated Scala file.