Update new-editor docs.
This updates the documentation for new clients and is focused on
configuring the client with `InitializationOptions` rather than
properties. All of the properties that are able to be set but are
also able to be set via `InitializationOptions` have been removed
from the documentation. The same for values under `experimental`,
which are now under `InitializationOptions`.